+++ Self_Driving_Car/CarND-Extended-Kalman-Filter-Project/src/tools.cpp
@@ -56,22 +56,22 @@ MatrixXd Tools::CalculateJacobian(const VectorXd& x_state) {
 
   MatrixXd Hj(3,4);
   //recover state parameters
-  double px = x_state(0);
-  double py = x_state(1);
-  double vx = x_state(2);
-  double vy = x_state(3);
+  float px = x_state(0);
+  float py = x_state(1);
+  float vx = x_state(2);
+  float vy = x_state(3);
 
   //Avoid zero in dr values
   if( fabs(px) < 0.001 ) px = 0.001;
   if( fabs(py) < 0.001 ) py = 0.001;
   
   //pre-compute a set of terms to avoid repeated calculation
-  double c1 = px*px+py*py;
+  float c1 = px*px+py*py;
   
   //c1 should not be zero
   if( fabs(c1) < 0.001 ) c1 = 0.001;
-  double c2 = sqrt(c1);
-  double c3 = (c1*c2);
+  float c2 = sqrt(c1);
+  float c3 = (c1*c2);
 
   //compute the Jacobian matrix
   Hj << (px/c2), (py/c2), 0.0, 0.0,
